fix: change button, code structure, change navigator
